+++ front/src/pages/portfolio/index.ts
@@ -1,1 +1,1 @@
-export { default } from '@routes/Portfolio';
+export { default } from '@pages/Portfolio';
